change(verification) reduce complexity by delete and rewrite methods
Delete methods isUserHaveRoleWithName isUserNotVerified and let handle it in orm and db side.

+++ src/main/java/com/epam/finalproject/service/impl/VerificationTokenServiceImpl.java
@@ -76,8 +76,11 @@ public class VerificationTokenServiceImpl implements VerificationTokenService {
     @Transactional
     @Loggable
     public void verifyByToken(VerificationToken token) {
-        Role role = roleRepository.findByName(RoleEnum.UNVERIFIED).orElseThrow();
+        if (isExpired(token)){
+            throw new IllegalArgumentException("Token expired");
+        }
         User user = token.getUser();
+        Role role = roleRepository.findByName(RoleEnum.UNVERIFIED).orElseThrow();
         user.deleteRole(role);
         userRepository.save(user);
         verificationTokenRepository.delete(token);
